showing the right star positions, note: on skycharts, left is east and right is west

+++ Classes/ALTimeManager.h
@@ -14,7 +14,7 @@
 	NSTimer* timeTest;
 	NSDate* simulatedDate;
 	NSDate* actualDate;
-	int speed;	
+	float speed;	
 }
 
 @property(readwrite, assign) NSDate* actualDate;
